# docs/openapi.yml
openapi: 3.1.0
info:
  title: Tunecamp REST API
  version: 2.0.0
  description: |
    The Tunecamp API provides programmatic access to the platform's music library, artists, albums, tracks,
    releases, playlists, and administrative functions. It supports the decentralized music ecosystem via
    GunDB and ActivityPub integrations.

    ## Authentication
    Public endpoints (catalog browsing, streaming) require no authentication.
    User endpoints require a Bearer token from `/api/auth/login`.
    Admin endpoints require a Bearer token with an admin role.

    ## Subsonic API
    For Subsonic/OpenSubsonic client compatibility, use the `/rest` endpoint.
    See [SUBSONIC.md](./SUBSONIC.md) for the full endpoint reference.
servers:
  - url: /api
    description: Tunecamp API base path
paths:
...
  /artists/{slug}/cover:
    get:
      summary: Get Artist Cover Image
      tags: [Artists]
      description: Returns the artist's profile photo/avatar.
      operationId: getArtistCover
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image file.

  # ── Albums ──────────────────────────────────────────────────────────────────
...
  /albums/{id}/cover:
    get:
      summary: Get Album Cover Art
      tags: [Albums]
      description: Returns the album's cover art image.
      operationId: getAlbumCover
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image file.

  # ── Tracks ──────────────────────────────────────────────────────────────────
...
  /tracks/{id}:
    get:
      summary: Get Track Details
      tags: [Tracks]
      description: Get a single track by ID.
      operationId: getTrack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Track details.
        '404':
          description: Track not found.
    put:
      summary: Update Track
      tags: [Tracks]
      description: Update track metadata (requires authentication, ownership check).
      operationId: updateTrack
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Track updated.
    delete:
      summary: Delete Track
      tags: [Tracks]
      description: Delete a track (requires authentication, ownership check).
      operationId: deleteTrack
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Track deleted.
...
  /playlists:
    get:
      summary: List Playlists
      tags: [Playlists]
      description: Retrieves all public playlists.
      operationId: getPlaylists
      responses:
        '200':
          description: A list of playlists.
    post:
      summary: Create Playlist
      tags: [Playlists]
      description: Create a new playlist (requires authentication).
      operationId: createPlaylist
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Playlist created.
  /playlists/{id}:
    get:
      summary: Get Playlist
      tags: [Playlists]
      description: Get a playlist with its tracks.
      operationId: getPlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Playlist with tracks.
    delete:
      summary: Delete Playlist
      tags: [Playlists]
      description: Delete a playlist (requires ownership).
      operationId: deletePlaylist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Playlist deleted.

  # ── Stats ───────────────────────────────────────────────────────────────────
...
  /catalog/search:
    get:
      summary: Search Catalog
      tags: [Search]
      description: Full-text search across artists, albums, and tracks.
      operationId: searchCatalog
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query.
      responses:
        '200':
          description: Search results grouped by type.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...